+++ src/tada-configs.ads
@@ -1,0 +1,9 @@
+package Tada.Configs is
+   Config_Error : exception;
+
+   type Config is record
+      Sections : Section_Maps.Map;
+   end record;
+
+   function Read (Config_Path : String) return Config;
+end Tada.Configs;